+++ src/aqua_diagnostics/global_biases/plot_global_biases.py
@@ -102,7 +102,7 @@ class PlotGlobalBiases:
         realization = get_realizations(data)
         proj = get_projection(proj, **proj_params)
         
-        title = (f"Climatology of {data[var].attrs.get('long_name', var)} \n for {data.AQUA_model} {data.AQUA_exp}" 
+        title = (f"Climatology of {data[var].attrs.get('long_name', var).lower()} \n for {data.AQUA_model} {data.AQUA_exp}" 
                 + (f" at {int(plev / 100)} hPa" if plev else ""))
 
         fig, ax = plot_single_map(
@@ -121,7 +121,7 @@ class PlotGlobalBiases:
         ax.set_ylabel("Latitude")
 
         description = (
-            f"Spatial map of the climatology for {data[var].attrs.get('long_name', var)}"
+            f"Spatial map of the climatology for {data[var].attrs.get('long_name', var).lower()}"
             f"{' at ' + str(int(plev / 100)) + ' hPa' if plev else ''}"
             f" from {time_to_string(data.startdate, format='%Y-%m')} to {time_to_string(data.enddate, format='%Y-%m')} for the {data.AQUA_model} model, experiment {data.AQUA_exp}."
         )
@@ -156,7 +156,7 @@ class PlotGlobalBiases:
 
         proj = get_projection(proj, **proj_params)
 
-        title = (f"Global difference of {data[var].attrs.get('long_name', var)} \n for {data.AQUA_model} {data.AQUA_exp}"
+        title = (f"Global difference of {data[var].attrs.get('long_name', var).lower()} \n for {data.AQUA_model} {data.AQUA_exp} "
                  f"relative to {data_ref.AQUA_model} climatology"
                  + (f" at {int(plev / 100)} hPa" if plev else ""))
 
@@ -180,9 +180,9 @@ class PlotGlobalBiases:
         
 
         description = (
-            f"Climatology of {data[var].attrs.get('long_name', var)}"
+            f"Climatology of {data[var].attrs.get('long_name', var).lower()}"
             f"{' at ' + str(int(plev / 100)) + ' hPa' if plev else ''} "
-            f"{data.AQUA_model} {data.AQUA_exp} (from {time_to_string(data.startdate, format='%Y-%m')} to {time_to_string(data.enddate, format='%Y-%m')}, contours) "
+            f"for {data.AQUA_model} {data.AQUA_exp} (from {time_to_string(data.startdate, format='%Y-%m')} to {time_to_string(data.enddate, format='%Y-%m')}, contours) "
             f"and differences against {data_ref.AQUA_model} (from {time_to_string(data_ref.startdate, format='%Y-%m')} to {time_to_string(data_ref.enddate, format='%Y-%m')}, shading)."
         )
 
@@ -218,7 +218,7 @@ class PlotGlobalBiases:
         season_list = ['DJF', 'MAM', 'JJA', 'SON']
         sym = vmin is None or vmax is None
 
-        title = (f"Seasonal difference of {data[var].attrs.get('long_name', var)} \n for {data.AQUA_model} {data.AQUA_exp} "
+        title = (f"Seasonal difference of {data[var].attrs.get('long_name', var).lower()} \n for {data.AQUA_model} {data.AQUA_exp} "
                  f"relative to {data_ref.AQUA_model} climatology"
                  + (f" at {int(plev / 100)} hPa" if plev else ""))
 
@@ -247,9 +247,9 @@ class PlotGlobalBiases:
 
 
         description = (
-            f"Seasonal climatology of {data[var].attrs.get('long_name', var)}"
+            f"Seasonal climatology of {data[var].attrs.get('long_name', var).lower()}"
             f"{' at ' + str(int(plev / 100)) + ' hPa' if plev else ''} "
-            f"{data.AQUA_model} {data.AQUA_exp} (from {time_to_string(data.startdate, format="%Y-%m")} to {time_to_string(data.enddate, format="%Y-%m")}, contours) "
+            f"for {data.AQUA_model} {data.AQUA_exp} (from {time_to_string(data.startdate, format="%Y-%m")} to {time_to_string(data.enddate, format="%Y-%m")}, contours) "
             f"and differences against {data_ref.AQUA_model} (from {time_to_string(data_ref.startdate, format="%Y-%m")} to {time_to_string(data_ref.enddate, format="%Y-%m")}, shading)."
         )
 
@@ -279,12 +279,12 @@ class PlotGlobalBiases:
         realization = get_realizations(data)
 
         title = (
-            f"Vertical difference of {data[var].attrs.get('long_name', var)} for \n {data.AQUA_model} {data.AQUA_exp}"
+            f"Vertical difference of {data[var].attrs.get('long_name', var).lower()} for \n {data.AQUA_model} {data.AQUA_exp} "
             f"relative to {data_ref.AQUA_model} climatology\n"
         )
 
         description = (
-            f"Vertical cross-section of {data[var].attrs.get('long_name', var)}"
+            f"Vertical cross-section of {data[var].attrs.get('long_name', var).lower()} for "
             f"{data.AQUA_model} {data.AQUA_exp} (from {time_to_string(data.startdate, format='%Y-%m')} to {time_to_string(data.enddate, format='%Y-%m')}, contours) "
             f"and differences against {data_ref.AQUA_model} (from {time_to_string(data_ref.startdate, format='%Y-%m')} to {time_to_string(data_ref.enddate, format='%Y-%m')}, shading)."
         )