+++ src/aqua_diagnostics/seaice/seaice.py
@@ -7,6 +7,7 @@ from aqua.logger import log_configure, log_history
 from aqua.util import to_list, merge_attrs
 from aqua.diagnostics.seaice.util import ensure_istype
 from aqua.fldstat import FldStat
+from aqua.util.time import time_to_string
 
 xr.set_options(keep_attrs=True)
 
@@ -514,8 +515,8 @@ class SeaIce(Diagnostic):
                                                  f"over {da_seaice_computed.attrs['AQUA_region']}")
         da_seaice_computed.attrs["standard_name"] = f"{region}_{'std_' if std_flag else ''}sea_ice_{self.method}"
         da_seaice_computed.attrs["AQUA_method"] = f"{self.method}"
-        if startdate is not None: da_seaice_computed.attrs["AQUA_startdate"] = f"{startdate}"
-        if enddate is not None: da_seaice_computed.attrs["AQUA_enddate"] = f"{enddate}"
+        if startdate is not None: da_seaice_computed.attrs["AQUA_startdate"] = time_to_string(startdate, format='%Y-%m')
+        if enddate is not None: da_seaice_computed.attrs["AQUA_enddate"] = time_to_string(enddate, format='%Y-%m')
         da_seaice_computed.name = f"{'std_' if std_flag else ''}sea_ice_{self.method}_{region}"
 
         time_coord = da_seaice_computed.coords.get("time", None)