+++ CHANGELOG.md
@@ -9,6 +9,7 @@ AQUA core complete list:
 - Netcdf4 and h5py in env instead of pip (#2739)
 
 AQUA diagnostics complete list:
+- ECmean: Enable time selection (#2761)
 
 ## [v0.19.10]